requirements.txt
pywebview>=3.7
